requirements.txt
-e .
pyusb>=1.0.0
